CDAP-19706 moved inlay drawer widget component out of common folder
