+++ app/cdap/components/InlayDrawerWidget/Menu.tsx
@@ -22,8 +22,8 @@ import MenuList from '@material-ui/core/MenuList';
 import Paper from '@material-ui/core/Paper';
 import Popper from '@material-ui/core/Popper';
 import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown';
-import { getTestIdString } from 'components/common/InlayDrawerWidget/index';
-import { PREFIX } from 'components/common/InlayDrawerWidget/InlayDrawerWidget.stories';
+import { getTestIdString } from 'components/InlayDrawerWidget/index';
+import { PREFIX } from 'components/InlayDrawerWidget/InlayDrawerWidget.stories';
 import T from 'i18n-react';
 import React, { KeyboardEvent, MouseEvent, useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
