+++ app/cdap/components/InlayDrawerWidget/index.tsx
@@ -17,7 +17,7 @@
 import { Box, IconButton, Typography } from '@material-ui/core';
 import grey from '@material-ui/core/colors/grey';
 import CloseRoundedIcon from '@material-ui/icons/CloseRounded';
-import Menu, { IActionsOptions } from 'components/common/InlayDrawerWidget/Menu';
+import Menu, { IActionsOptions } from 'components/InlayDrawerWidget/Menu';
 import React, { PropsWithChildren } from 'react';
 import styled from 'styled-components';
